requirements.txt
faiss-gpu==1.7.2
numpy==1.22.3
pandas==1.4.3
tqdm==4.64.0
transformers==4.20.1